+++ NEWS.md
@@ -3,6 +3,7 @@
 ----------------------------------
 
 # Version 2.4.1
+- Compliance with medfate v. 4.3.1
 - New helper function add_forests() to parse forest plot data tables
 - New option `simplify` in function initialize_landscape()